Add format mode and make soft mode more loose
Previously, oddly formatted SSNs such as 12-123-1234 would not be
allowed in soft/loose mode.

This patch corrects this behaviour so that any combination of dashes or
spaces between those digits is allowed in soft mode. The only requirement is that the
resulting number combination is actually a valid SSN.

The previous behaviour can be achieved by using format mode.